+++ src/main/java/mod/gottsch/neo/claimmyland/core/item/ModItems.java
@@ -107,6 +107,13 @@ public class ModItems {
 
     // tools
 //    public static DeferredHolder<Item> ZONING_TOOL = Registration.ITEMS.register("zoning_tool", () -> new ZoningTool(new Item.Properties()));
+    public static final DeferredHolder<Item, IronNameTagItem> IRON_NAME_TAG =
+            Registration.ITEMS.register("iron_name_tag",
+                    () -> new IronNameTagItem(new Item.Properties().stacksTo(16)));
+
+    public static final DeferredHolder<Item, GoldNameTagItem> GOLD_NAME_TAG =
+            Registration.ITEMS.register("gold_name_tag",
+                    () -> new GoldNameTagItem(new Item.Properties().stacksTo(16)));
 
     /**
      *